split repositories by responsibility and extract auth/location/user infrastructure
