+++ domain/src/main/java/org/monogram/domain/repository/LocationRepository.kt
@@ -4,5 +4,4 @@ import org.monogram.domain.models.webapp.OSMReverseResponse
 
 interface LocationRepository {
     suspend fun reverseGeocode(lat: Double, lon: Double): OSMReverseResponse?
-    suspend fun searchLocation(query: String): List<OSMReverseResponse>
 }
