+++ domain/src/main/java/org/monogram/domain/repository/SponsorRepository.kt
@@ -1,0 +1,5 @@
+package org.monogram.domain.repository
+
+interface SponsorRepository {
+    fun forceSponsorSync()
+}
