+++ presentation/src/main/java/org/monogram/presentation/features/profile/admin/DefaultMemberListComponent.kt
@@ -7,9 +7,9 @@ import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import org.monogram.domain.models.GroupMemberModel
+import org.monogram.domain.repository.ChatInfoRepository
 import org.monogram.domain.repository.ChatMemberStatus
 import org.monogram.domain.repository.ChatMembersFilter
-import org.monogram.domain.repository.UserRepository
 import org.monogram.presentation.core.util.componentScope
 import org.monogram.presentation.root.AppComponentContext
 
@@ -22,7 +22,7 @@ class DefaultMemberListComponent(
     private val onMemberLongClicked: (Long) -> Unit
 ) : MemberListComponent, AppComponentContext by context {
 
-    private val userRepository: UserRepository = container.repositories.userRepository
+    private val chatInfoRepository: ChatInfoRepository = container.repositories.chatInfoRepository
 
     private val scope = componentScope
     private val _state = MutableValue(MemberListComponent.State(chatId = chatId, type = type))
@@ -48,7 +48,7 @@ class DefaultMemberListComponent(
                     MemberListComponent.MemberListType.BLACKLIST -> ChatMembersFilter.Banned
                 }
 
-                val members = userRepository.getChatMembers(chatId, offset, limit, filter)
+                val members = chatInfoRepository.getChatMembers(chatId, offset, limit, filter)
 
                 if (members.isEmpty()) {
                     _state.update { it.copy(canLoadMore = false) }
@@ -101,7 +101,7 @@ class DefaultMemberListComponent(
                 _state.update { it.copy(isLoading = true) }
                 try {
                     val filter = ChatMembersFilter.Search(query)
-                    val results = userRepository.getChatMembers(chatId, 0, 50, filter)
+                    val results = chatInfoRepository.getChatMembers(chatId, 0, 50, filter)
 
                     val filtered = when (type) {
                         MemberListComponent.MemberListType.ADMINS -> results.filter { it.status is ChatMemberStatus.Administrator || it.status is ChatMemberStatus.Creator }
